+++ src/main/java/com/btc/snow/admin/assignment/AdminAssignmentController.java
@@ -1,6 +1,5 @@
 package com.btc.snow.admin.assignment;
 
-import com.btc.snow.admin.config.UploadFileServiceForAdmin;
 import com.btc.snow.admin.member.AdminMemberDto;
 import com.btc.snow.user.assignment.UserAssignmentDto;
 import com.btc.snow.user.member.UserMemberDto;
