+++ src/main/java/com/btc/snow/admin/assignment/DownloadFileControllerForAdmin.java
@@ -1,4 +1,4 @@
-package com.btc.snow.user.config;
+package com.btc.snow.admin.assignment;
 
 import lombok.extern.log4j.Log4j2;
 import org.springframework.core.io.Resource;
@@ -17,38 +17,34 @@ import java.nio.file.Paths;
 
 @Log4j2
 @Controller
-public class DownloadFileControllerForUser {
+public class DownloadFileControllerForAdmin {
+    private static final String DOWNLOAD_DIR = "c:/snow/admin/assignment"; // 파일이 저장된 경로
 
-    private static final String DOWNLOAD_DIR = "c:/snow/user/assignment";
-
-    @GetMapping("/downloadFromUser")
-    public ResponseEntity<Resource> downloadFile(@RequestParam String file, @RequestParam String user_id) {
-        log.info("[DownloadFileControllerForUser] downloadFile");
-        log.info("downloadFromUser file :::::::::::: " + file);
-        log.info("downloadFromUser user_id  :::::::::::: " + user_id);
-
-//        String download_dir = "c:/snow/user/assignment" + user_id;
+    @GetMapping("/downloadFromAdmin")
+    public ResponseEntity<Resource> downloadFile(@RequestParam String file) {
+        log.info("[DownloadFileController] downloadFile");
+        log.info("download file name :::::::::::: " + file);
 
         try {
             // 파일 경로 생성
-            Path filePath = Paths.get(DOWNLOAD_DIR + "/" + user_id).resolve(file).normalize();
+            Path filePath = Paths.get(DOWNLOAD_DIR).resolve(file).normalize();
             Resource resource = new UrlResource(filePath.toUri());
 
             if (resource.exists() && resource.isReadable()) {
                 // 파일이 존재하고 읽을 수 있는 경우 다운로드 응답 생성
-                log.info("[DownloadFileControllerForUser] 파일 존재");
+                log.info("[DownloadFileControllerForAdmin] 파일 존재");
                 return ResponseEntity.ok()
                         .contentType(MediaType.APPLICATION_OCTET_STREAM)
                         .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; file=\"" + resource.getFile() + "\"")
                         .body(resource);
             } else {
                 // 파일이 존재하지 않거나 읽을 수 없는 경우 404 에러 응답
-                log.info("[DownloadFileControllerForUser] 파일 없음");
+                log.info("[DownloadFileControllerForAdmin] 파일 없음");
                 return ResponseEntity.notFound().build();
             }
         } catch (MalformedURLException e) {
             // 파일 경로 변환 실패 시 400 에러 응답
-            log.info("[DownloadFileControllerForUser] 파일 경로 변환 실패");
+            log.info("[DownloadFileControllerForAdmin] 파일 경로 변환 실패");
             return ResponseEntity.badRequest().build();
 
         } catch (IOException e) {
