+++ src/main/java/com/btc/snow/admin/assignment/UploadFileServiceForAdmin.java
@@ -1,4 +1,4 @@
-package com.btc.snow.user.config;
+package com.btc.snow.admin.assignment;
 
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
@@ -9,10 +9,10 @@ import java.util.UUID;
 
 @Log4j2
 @Service
-public class UploadFileServiceForUser {
+public class UploadFileServiceForAdmin {
 
-    public String upload(MultipartFile file, String user_id) {
-        log.info("[UploadFileServiceForUser] upload()");
+    public String upload(MultipartFile file) {
+        log.info("[UploadFileServiceForAdmin] upload()");
 
         // 저장 유무
         boolean result = false;
@@ -20,7 +20,7 @@ public class UploadFileServiceForUser {
         // File 저장
         String fileOriName = file.getOriginalFilename();
         String fileExtension = fileOriName.substring(fileOriName.lastIndexOf("."), fileOriName.length());
-        String uploadDir = "c:\\snow\\user\\assignment\\" + user_id; // 저장할 주소 !! user_id 별로 받아야지
+        String uploadDir = "c:\\snow\\admin\\assignment";
 
         UUID uuid = UUID.randomUUID(); // 랜덤으로 유니크한 값을 생성해서 보내주는 메소드 (덮어쓰기 방지)
         String uniqueName = uuid.toString().replace("-", ""); // 값이 000-000-.. 이런형식이라서 걍 "-" 없애기 [선택사항]
@@ -42,12 +42,12 @@ public class UploadFileServiceForUser {
 
         // 성공시 저장
         if (result == true) {
-            log.info("[UploadFileServiceForUser] FILE UPLOAD SUCCESS!!");
+            log.info("[UploadFileServiceForAdmin] FILE UPLOAD SUCCESS!!");
 
             return uniqueName + fileExtension;
 
         } else {
-            log.info("[UploadFileServiceForUser] FILE UPLOAD FAIL!!");
+            log.info("[UploadFileServiceForAdmin] FILE UPLOAD FAIL!!");
 
             return null;
