+++ src/main/java/com/btc/snow/admin/attendence/UploadFileService.java
@@ -1,4 +1,4 @@
-package com.btc.snow.admin.config;
+package com.btc.snow.admin.attendence;
 
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
@@ -9,10 +9,10 @@ import java.util.UUID;
 
 @Log4j2
 @Service
-public class UploadFileServiceForAdmin {
+public class UploadFileService {
 
     public String upload(MultipartFile file) {
-        log.info("[UploadFileServiceForAdmin] upload()");
+        log.info("[UploadFileService] upload()");
 
         // 저장 유무
         boolean result = false;
@@ -20,7 +20,8 @@ public class UploadFileServiceForAdmin {
         // File 저장
         String fileOriName = file.getOriginalFilename();
         String fileExtension = fileOriName.substring(fileOriName.lastIndexOf("."), fileOriName.length());
-        String uploadDir = "c:\\snow\\admin\\assignment";
+        String uploadDir = "c:\\snow/user/attendence";
+//        String uploadDir = "c:\\snow\\member\\" + user_id; // 아이디별로 받겠다 여기에 키득
 
         UUID uuid = UUID.randomUUID(); // 랜덤으로 유니크한 값을 생성해서 보내주는 메소드 (덮어쓰기 방지)
         String uniqueName = uuid.toString().replace("-", ""); // 값이 000-000-.. 이런형식이라서 걍 "-" 없애기 [선택사항]
@@ -32,7 +33,7 @@ public class UploadFileServiceForAdmin {
 
         try {
             file.transferTo(saveFile);
-            
+
             result = true;
 
         } catch (Exception e) {
@@ -42,12 +43,12 @@ public class UploadFileServiceForAdmin {
 
         // 성공시 저장
         if (result == true) {
-            log.info("[UploadFileServiceForAdmin] FILE UPLOAD SUCCESS!!");
+            log.info("[UploadFileService] FILE UPLOAD SUCCESS!!");
 
             return uniqueName + fileExtension;
 
         } else {
-            log.info("[UploadFileServiceForAdmin] FILE UPLOAD FAIL!!");
+            log.info("[UploadFileService] FILE UPLOAD FAIL!!");
 
             return null;
